the pianist

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,15 +9,15 @@
 int main(void)
 {
 	int i = 0;
-	long int a = 0, b = 1;
+	long long int a = 0, b = 1;
 
 	for (; i < 97; i++)
 	{
 		b += a;
 		a = b - a;
-		printf("%lu, ", b);
+		printf("%llu, ", b);
 	}
-	printf("%lu\n", a + b);
+	printf("%llu\n", a + b);
 
 	return (0);
 }
